WIP: move to having a single test suite per application
