+++ java/java-node/gradle_node_test.go
@@ -1,4 +1,4 @@
-package java_test
+package java_node_test
 
 import (
 	"flag"
@@ -23,21 +23,21 @@ func init() {
 	flag.Var(&builders, "name", "the name a builder to test with")
 }
 
-func TestJava(t *testing.T) {
+func TestGradleNode(t *testing.T) {
 	Expect := NewWithT(t).Expect
 
 	Expect(len(builders)).NotTo(Equal(0))
 
 	SetDefaultEventuallyTimeout(60 * time.Second)
 
-	suite := spec.New("Java", spec.Parallel(), spec.Report(report.Terminal{}))
+	suite := spec.New("Java - Gradle & Node", spec.Parallel(), spec.Report(report.Terminal{}))
 	for _, builder := range builders {
-		suite(fmt.Sprintf("Java with %s builder", builder), testJavaWithBuilder(builder), spec.Sequential())
+		suite(fmt.Sprintf("Gradle Node with %s builder", builder), testGradleNodeWithBuilder(builder), spec.Sequential())
 	}
 	suite.Run(t)
 }
 
-func testJavaWithBuilder(builder string) func(*testing.T, spec.G, spec.S) {
+func testGradleNodeWithBuilder(builder string) func(*testing.T, spec.G, spec.S) {
 	return func(t *testing.T, context spec.G, it spec.S) {
 		var (
 			Expect     = NewWithT(t).Expect
@@ -86,48 +86,6 @@ func testJavaWithBuilder(builder string) func(*testing.T, spec.G, spec.S) {
 				}
 
 				Expect(os.RemoveAll(source)).To(Succeed())
-			})
-
-			context("app uses maven and yarn", func() {
-				it("builds successfully", func() {
-					var err error
-					source, err = occam.Source(filepath.Join(".", "maven-yarn"))
-					Expect(err).NotTo(HaveOccurred())
-
-					var logs fmt.Stringer
-					image, logs, err = pack.Build.
-						WithPullPolicy("never").
-						WithBuilder(builder).
-						WithVolumes(fmt.Sprintf("%s/.m2:/home/cnb/.m2:rw", home)).
-						WithGID("123").
-						WithEnv(map[string]string{
-							"BP_JVM_VERSION":       "17",
-							"BP_JAVA_INSTALL_NODE": "true"}).
-						Execute(name, source)
-
-					Expect(err).ToNot(HaveOccurred(), logs.String)
-
-					Expect(logs).To(ContainLines(ContainSubstring("Paketo Buildpack for CA Certificates")))
-					Expect(logs).To(ContainLines(ContainSubstring("Paketo Buildpack for BellSoft Liberica")))
-					Expect(logs).To(ContainLines(ContainSubstring("Paketo Buildpack for Yarn")))
-					Expect(logs).To(ContainLines(ContainSubstring("Paketo Buildpack for Node Engine")))
-					Expect(logs).To(ContainLines(ContainSubstring("Paketo Buildpack for Maven")))
-					Expect(logs).To(ContainLines(ContainSubstring("Paketo Buildpack for Executable JAR")))
-					Expect(logs).To(ContainLines(ContainSubstring("Paketo Buildpack for Spring Boot")))
-
-					container, err = docker.Container.Run.
-						WithPublish("8080").
-						Execute(image.ID)
-					Expect(err).NotTo(HaveOccurred())
-
-					Eventually(container).Should(Serve(ContainSubstring("UP")).OnPort(8080).WithEndpoint("/actuator/health"))
-					Eventually(container).
-						Should(
-							Serve(MatchRegexp("<script type=\"module\" crossorigin src=\"/assets/index-.*.js\"></script>")).
-								OnPort(8080).
-								WithEndpoint("/"))
-
-				})
 			})
 
 			context("app uses gradle and node", func() {
